requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
openai==1.51.0
pinecone-client==5.0.1
playwright==1.50.0